HDF5-based parallel IO preparation
This patch migrates the parallel netcdf file creation to use the
HDF5-based parallel library, rather than the pnetCDF implementation.

This required that variables be redefined so that writes on time-based
variables are now collective.

Parallel netcdf is now enabled using the mpp_io namelist variable
`parallel_netcdf`.  This should eliminate the need to use the
`PAR_ACCESS` directive.  While this flag is not yet completely safe, it
should eventually be safe to implement this without a preprocessor
directive when netCDF4 is enabled, since the netCDF4 library always has
implementations of the parallel libraries, even if they don't do
anything.

Also, the `parallel_netcdf` flag should probably eventually be moved to
the `fms_io` module, where most netCDF control flags seem to be saved.

This is not yet a proper implementation; a full copy of the fields are
saved to each rank, and then all ranks write identical data to the file.
This will probably fail for very large CPU counts, where
`mpp_global_field` tends to push against the limits of the MPI
libraries.

A proper implementation will set each rank to only write its
own local domain to the file.  The purpose of this patch is only to get
a working version.