parallel_namelist flag bugfix (no PAR_ACCESS)
File creation and opening was still within a hard-compiled preprocessor
(`PAR_ACCESS`) which was hanging when `parallel_netcdf` was unset.

There is now a `parallel_netcdf` flag which determines if the (blocking)
parallel open/creation functions are called.

(I have yet to test this in a serial netcdf library, but it should be ok
...?)